+++ crates/backend-uzu/src/backends/metal/kernel/matmul/gemm/common/quant_scale_zero_point.h
@@ -79,9 +79,9 @@ struct QuantizedBlockLoaderScaleZeroPoint {
         zero_points_row_start(
             SCALE_SYMMETRIC
                 ? nullptr
-                : (REDUCTION_DIMENSION == 1
-                       ? (zero_points_row_start_ + tile_row_index * zero_point_row_stride<BITS>(groups_per_row_))
-                       : zero_points_row_start_)
+                : (REDUCTION_DIMENSION == 1 ? (zero_points_row_start_ +
+                                               tile_row_index * zero_point_row_stride<ushort(BITS)>(groups_per_row_))
+                                            : zero_points_row_start_)
         ) {}
 
   inline void current_scale_bias(thread T& out_scale, thread T& out_bias) const {
@@ -96,9 +96,9 @@ struct QuantizedBlockLoaderScaleZeroPoint {
       scale_value = *scales;
     }
     if constexpr (SCALE_SYMMETRIC) {
-      zero_point_value = symmetric_zero_point<BITS>();
+      zero_point_value = symmetric_zero_point<ushort(BITS)>();
     } else {
-      zero_point_value = decode_zero_point<BITS>(zero_points_row_start, uint(group_index));
+      zero_point_value = decode_zero_point<ushort(BITS)>(zero_points_row_start, uint(group_index));
     }
     out_scale = scale_value;
     out_bias = static_cast<T>(-scale_value * static_cast<T>(zero_point_value));
